added dll implementation and edited exe output
