+++ README.md
@@ -2,19 +2,20 @@
 
 ![WTFPL badge](http://www.wtfpl.net/wp-content/uploads/2012/12/wtfpl-badge-4.png)
 
-isWINE is a very small and simple application to check if your app is running in a Linux WINE environment.
+isWINE is a very small and simple application and dynamic link library to check if your app is running in a Linux WINE environment.
 
 #### Output on Windows:
 
 ```
-> isWINE.exe
+$> isWINE.exe
 0
 ```
 
 #### Output on Linux:
 
 ```bash
-$ wine isWINE.exe 
+$> wine isWINE.exe
+1
 7.0
 ```
