--- isWINE/isWINE.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32413.511
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "isWINE", "isWINE.vcxproj", "{87E8A266-6086-4402-B0BC-54B38E6BE028}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Debug|x64.ActiveCfg = Debug|x64
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Debug|x64.Build.0 = Debug|x64
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Debug|x86.ActiveCfg = Debug|Win32
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Debug|x86.Build.0 = Debug|Win32
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Release|x64.ActiveCfg = Release|x64
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Release|x64.Build.0 = Release|x64
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Release|x86.ActiveCfg = Release|Win32
-		{87E8A266-6086-4402-B0BC-54B38E6BE028}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1928B4B3-D9AD-46CE-B8EB-F7DAEB029FE2}
-	EndGlobalSection
-EndGlobal
